+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to
 
 ## [Unreleased]
 
+## [v0.3.0] - 2026-09-17
+
 ### Added
 
 - Implement service provider credentials automated generation, encryption and
@@ -26,7 +28,7 @@ Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to
 
 - Clean unused `django-cors-headers` configured dependency
 
-## [0.2.0] - 2026-07-29
+## [v0.2.0] - 2026-07-29
 
 ### Changed
 
@@ -41,12 +43,13 @@ Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to
 - Upgrade `sentry-sdk` to `2.66.1`
 - Upgrade `uvicorn` to `0.51.0`
 
-## [0.1.0] - 2026-07-14
+## [v0.1.0] - 2026-07-14
 
 ### Added
 
 - Implement base Token Exchange (RFC 8693) endpoints
 
-[unreleased]: https://github.com/suitenumerique/menshen/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/suitenumerique/menshen/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/suitenumerique/menshen/releases/v0.3.0
 [v0.2.0]: https://github.com/suitenumerique/menshen/releases/v0.2.0
 [v0.1.0]: https://github.com/suitenumerique/menshen/releases/v0.1.0